review handling of data reader line to string[]
